requirements.txt
# requirements.txt
# ConstructFlix - Enterprise Construction Management Platform
# Python 3.11+ required

# ========== CORE FRAMEWORK ==========
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Werkzeug==2.3.7
WTForms==3.1.0
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2

# ========== DATABASE ==========
psycopg2-binary==2.9.7  # PostgreSQL
SQLAlchemy==2.0.21
alembic==1.12.0
redis==5.0.0  # Redis cache

# ========== AUTHENTICATION & SECURITY ==========
python-jose[cryptography]==3.3.0  # JWT tokens
passlib==1.7.4  # Password hashing
bcrypt==4.0.1  # Password hashing
Flask-Bcrypt==1.0.1
email-validator==2.1.0
PyJWT==2.8.0
cryptography==41.0.5  # Encryption
python-dotenv==1.0.0  # Environment variables

# ========== API & SERIALIZATION ==========
Flask-RESTful==0.3.10
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
webargs==8.2.0
apispec==6.3.0
flask-swagger-ui==4.11.1

# ========== REAL-TIME COMMUNICATION ==========
Flask-SocketIO==5.3.4
python-socketio==5.9.0
eventlet==0.33.3  # For WebSocket production
gevent==23.9.1  # Alternative async
gevent-websocket==0.10.1

# ========== SCHEDULING & ALGORITHMS ==========
networkx==3.1  # Critical path analysis
pandas==2.1.0  # Data manipulation
numpy==1.24.3  # Numerical computations
plotly==5.17.0  # Gantt charts
matplotlib==3.7.2  # Charts
seaborn==0.12.2  # Statistical visualization
python-dateutil==2.8.2
pytz==2023.3

# ========== FILE PROCESSING ==========
openpyxl==3.1.2  # Excel files
xlrd==2.0.1  # Read Excel
xlwt==1.3.0  # Write Excel
python-magic==0.4.27  # File type detection
Pillow==10.0.1  # Image processing
reportlab==4.0.4  # PDF generation
pdfkit==1.0.0  # HTML to PDF
WeasyPrint==59.0  # Alternative PDF generation

# ========== EMAIL & NOTIFICATIONS ==========
Flask-Mail==0.9.1
sendgrid==6.10.0  # SendGrid integration
twilio==8.10.0  # SMS notifications
celery==5.3.4  # Async task queue
redis==5.0.0  # Celery broker
flower==2.0.0  # Celery monitoring

# ========== DATA IMPORT/EXPORT ==========
csvkit==1.2.0
tablib==3.5.0
xlrd==2.0.1
XlsxWriter==3.1.2
pyexcel==0.7.0
pyexcel-xls==0.7.0
pyexcel-xlsx==0.6.0
xmltodict==0.13.0
dicttoxml==1.7.16

# ========== CLOUD STORAGE ==========
boto3==1.28.62  # AWS S3
botocore==1.31.62
s3transfer==0.7.0
minio==7.1.14  # MinIO/S3 compatible
google-cloud-storage==2.10.0  # Google Cloud Storage
azure-storage-blob==12.17.0  # Azure Blob Storage

# ========== MONITORING & LOGGING ==========
prometheus-flask-exporter==0.22.4
prometheus-client==0.18.0
python-json-logger==2.0.7
loguru==0.7.2
elasticsearch==8.10.0  # ELK Stack
elastic-apm==6.15.1  # Elastic APM
sentry-sdk[flask]==1.31.0  # Error tracking
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-flask==0.41b0

# ========== PERFORMANCE ==========
Flask-Caching==2.1.0
Flask-Limiter==3.5.0  # Rate limiting
limits==3.6.0
python-memcached==1.59  # Alternative cache
gunicorn==21.2.0  # Production WSGI server
gevent==23.9.1
uvicorn[standard]==0.23.2  # ASGI server

# ========== TESTING ==========
pytest==7.4.2
pytest-cov==4.1.0
pytest-flask==1.2.0
pytest-mock==3.11.1
factory-boy==3.3.0
faker==19.6.1
locust==2.15.1  # Load testing
coverage==7.3.1
tox==4.11.3

# ========== DEVELOPMENT TOOLS ==========
black==23.10.0
flake8==6.1.0
pylint==3.0.1
mypy==1.5.1
pre-commit==3.4.0
isort==5.12.0
bandit==1.7.5  # Security linter
safety==2.3.5  # Dependency vulnerability check

# ========== UTILITIES ==========
requests==2.31.0
urllib3==2.0.6
certifi==2023.7.22
idna==3.4
chardet==5.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
tqdm==4.66.1
colorama==0.4.6
tabulate==0.9.0
humanize==4.8.0
inflection==0.5.1
python-slugify==8.0.1
unidecode==1.3.6